src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "fastmcp>=2.12.4",
    "trimble-id>=0.1.1",
    "uvicorn[standard]>=0.37.0",
]

[tool.uv.sources]
helphub = { path = "../../helphubp/helphub_cli/dist/helphub-0.1.0-py3-none-any.whl" }

[dependency-groups]
dev = [
    "helphub",
]
